requirements.txt
streamlit
streamlit-option-menu
pandas
numpy
matplotlib
seaborn
requests
python-dotenv
plotly
scikit-learn
scipy
beautifulsoup4
lxml
python-dateutil
tqdm
Pillow